Storage of memory body text moved from db to files
Database now containers only each memory's title and keywords in order to perform string matching. The body text for each memory is now in a separate data directory. Body content is retrieved by accessing the file whose name corresponds to the db_key of the database record for that memory.

+++ src/io.py
@@ -8,40 +8,28 @@ class DB:
         self.connection = sqlite3.connect(file_path)
         self.cursor = self.connection.cursor()
         self.cursor.execute("""CREATE TABLE IF NOT EXISTS records
-                            (key INTEGER, title TEXT, keywords TEXT,
-                            body TEXT, PRIMARY KEY (key))""")
+                            (key INTEGER, title TEXT, keywords TEXT, PRIMARY KEY (key))""")
+
     def dump(self):
         sql = """SELECT * FROM records"""
         self.cursor.execute(sql)
         return self.cursor.fetchall()
-    def insert(self, title, keywords, body):
-        sql = """INSERT INTO records(title, keywords, body) VALUES(?,?,?)"""
-        self.cursor.execute(sql, (title, keywords, body))
-        self.connection.commit()
-    def update(self, mem_obj, mem_ui):
-        """
-        :type mem_obj: memory.Memory
-        :type mem_ui: memory.UI
-        Updates to database on done if mem_obj has different values compared to mem_ui
-        """
-        sql = """UPDATE records SET"""
-        if mem_obj.title != mem_ui.title_text:
-            sql += """ title="{}",""".format(mem_ui.title_text)
-        if mem_obj.keywords != mem_ui.keywords_text:
-            sql += """ keywords="{}",""".format(mem_ui.keywords_text)
-        if mem_obj.body != mem_ui.body_text:
-            sql += """ body="{}",""".format(mem_ui.body_text)
 
-        # sql string will only change from default value if change has been detected
-        if sql != """UPDATE records SET""":
-            if sql.endswith(','):
-                sql = sql.rsplit(',',1)[0]
-            sql += """ WHERE key={}""".format(mem_obj.db_key)
-            self.cursor.executescript(sql)
-            self.connection.commit()
+    def insert(self, Mem):
+        sql = """INSERT INTO records(title, keywords) VALUES(?,?)"""
+        self.cursor.execute(sql, (Mem.title, Mem.keywords))
+        self.connection.commit()
+        return self.cursor.lastrowid
+
     def remove(self, key):
         sql = """DELETE FROM records WHERE key={}""".format(key)
         self.cursor.execute(sql)
+        self.connection.commit()
+
+    def update_many(self, record_key, record_changes={}):
+        cols_vals = ''.join(['{}="{}", '.format(col,val) for col,val in record_changes.items()])
+        sql = """UPDATE records SET {} WHERE key={key)""".format(cols_vals, key=record_key)
+        self.cursor.executescript(sql)
         self.connection.commit()
 
 def file_exists(filepath):
@@ -49,6 +37,16 @@ def file_exists(filepath):
         return True
     print('Invalid file path {}'.format(filepath))
     return False
+
+def delete_file(filepath):
+    """
+    :type filepath: str
+    """
+    try:
+        if file_exists(filepath):
+            os.remove(filepath)
+    except Exception as e:
+        print('Error occured while deleting {}\n{}'.format(filepath, e.args))
 
 def json_from_file(filepath):
     """
@@ -74,12 +72,23 @@ def json_to_file(filepath, payload):
     except Exception as e:
         print('Error occured while writing json to {}\n{}'.format(filepath, e.args))
 
-def mkfile(filepath):
+def init_dir(dirpath):
+    """
+    :type dirpath: str
+    """
+    try:
+        if not os.path.isdir(dirpath):
+            os.mkdir(dirpath)
+    except Exception as e:
+        print('Error occured while creating directory at {}\n{}'.format(dirpath, e.args))
+
+def init_file(filepath):
     """
     :type filepath: str
     """
     try:
-        open(filepath, 'w').close()
+        if not os.path.isfile(filepath):
+            open(filepath, 'w').close()
     except Exception as e:
         print('Error occured while making file {}\n{}'.format(filepath, e.args))
 
@@ -93,5 +102,29 @@ def set_from_file(filepath):
             with open(filepath, 'r') as f:
                 return set([line.strip() for line in f.readlines()])
         return set()
+    except Exception as e:
+        print('Error occured while reading {} as set\n{}'.format(filepath, e.args))
+
+def str_from_file(filepath):
+    """
+    :type filepath: str
+    :returns: contents of file at file_path as string
+    """
+    try:
+        if file_exists(filepath):
+            with open(filepath, 'r') as f:
+                return f.read()
+        return str()
+    except Exception as e:
+        print('Error occured while reading {} as string\n{}'.format(filepath, e.args))
+
+def str_to_file(filepath, payload):
+    """
+    :type filepath: str
+    :type payload: str
+    """
+    try:
+        with open(filepath, 'w') as f:
+            f.write(payload)
     except Exception as e:
         print('Error occured while reading {} as set\n{}'.format(filepath, e.args))
